requirements.txt
PyQt5>=5.15.0
Pillow>=9.0.0
pytesseract>=0.3.10
kivy>=2.3.0
SpeechRecognition>=3.8.1
pyttsx3>=2.90
pyaudio>=0.2.13
python-docx>=0.8.11
PyPDF2>=3.0.0
google-generativeai>=0.3.0
